pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "forgekeeper"
version = "0.0.0.dev0"
description = "Forgekeeper agent (mono-repo dev install)"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Proprietary" }
authors = [{ name = "Forgekeeper Team" }]
dependencies = [
  "GitPython>=3.1.40",
  "PyYAML>=6.0",
  "openai>=1.0.0",
  "paho-mqtt>=2.0.0",
  "python-dotenv>=1.0.0",
  "requests>=2.31.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0"]
transformers = ["transformers>=4.44.0", "torch"]
triton = ["tritonllm>=0.1.0", "tritonclient[http]>=2.45.0"]

[project.scripts]
forgekeeper = "forgekeeper.__main__:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["forgekeeper*"]
exclude = [
  "forgekeeper.backend*",
  "forgekeeper.frontend*",
  "forgekeeper_v1*",
]

